backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
sqlalchemy==2.0.36
passlib[bcrypt]==1.7.4
pyjwt==2.10.1
python-multipart==0.0.9
httpx==0.28.1
pydantic-settings==2.7.0
python-dotenv==1.0.1
